UserName Availability check on Registration Page
My first checkin to this project, this is a super small update.  I added
a new CSS file called forms which is being used to stored
success/failure validation stuff.  There is one function in the
regisration.cshtml page that should be moved out to a more global file
but didn't want to jump out on a limb.