Move the DB  seed code so it works in desktop env..
The automatic execution of the migration of the database needs to
execute on a 'clean' new database in development environments. That's
because VS now initializes the db as part of its build, so the project's
first run has a new correct DB with no data. That meant the cluster
objects weren't being created.