Introduced Bundles for CSS and JavaScript to reduce requests.
By using bundling the minified version of the stylesheet is kept in sync
with the original and http traffic between client and server is reduced.
Also moved the styles in the files Body.css and Forms.css and the
cluster-coordinator class (probably accidently placed in the file
bootstrap-theme instead of in a site specific one) to a newly created
file site.css to keep them together.