Fixed validation for User name when registering.
Delayed the check for availability of a username until three characters
are typed to prevent a message that the username is available when only
one or two characters are filled in.
The DataAnnotations on the viewmodel to check for the length of username
were on the private field instead of the public property so they were
never used.